web-platform/ai-services/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Core
sentence-transformers==2.2.2
torch==2.1.0

# Vector Database
qdrant-client==1.7.0

# Database & Storage
supabase==2.0.3
python-dotenv==1.0.0

# Utilities
numpy==1.24.3
pydantic==2.5.0
